infrastructure for editing sport topics
Still have not seen onChanged on storage ever been fired.
I'm not sure about the order of execution of various parts of
an extension. But I think the order is:

1. content_scripts
2. page_action

The two modules does not share realm, so there can not be any code sharing.
Rather, shared code needs to be imported twice and will run in separate
instances.